Add rotation scan widget, controller, and auxilliary fixes
Add rotation scan widget with controller, make changes in rotation widget and rotation controller, as well as StandaMotorManager, to make the additions work. A rotation scan can now be connected to the positioner scan. Bug fixing needed. Next step: find a way to connect the steps taken in the rotation scan with the actual frames taken during the scanning.