Add ScanWidget types and add scan length check before generation
Add different ScanWidget, ScanManager, and ScanController types, depending on the microscopy technique (Base and PointScan for now, add more later). Additionally add a check in GalvoScanDesigner to check if the curves that will be generated will be "too long" for the RAM to handle (rough check), and stop generation if so.