Add mockers and fix loading of LeicaDMIManager
Loading of LeicaDMIManager is now performed in the added StandManager (that could also support other microscope stands in the future). A mocker is loaded if the LeicaDMIManager is not found (due to NDA it is not provided). A SetupInfo class is added for a generic microscope stand with a rs232 connection.