Add saving/loading of coordinate transformation coefficients
Add saving and loading of coordinate transformation coefficients to/from csv files, allowing to quickly recover a previous coordinate transform calibration. Csv files are saved in \ImSwitchConfig\imcontrol_etsted\transform_pipelines, and all .csv files in this folder are listed in the dropdown menu for picking the coefficients of choice.