Make the SLMs Controller more relaxed, warnings instead of errors
Changed the strictness of the SLMs: e.g. if the correction pattern of the desired wavelength is not found give a warning and take the one that is closest. Similar for the wavetable file (=SLM grey value calibration). Specifying it as string called "none" for when no wavetable file  is not a great solution, but at least consistent with how configurations  are specified for other hardware, e.g. aaatof.